refactor: refatora codigo dos componentes
