+++ client/src/components/Card/EditCardModal/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { Container } from './styles';
+
+export const EditCardModal: React.FC = () => {
+  return <Container />;
+};
