+++ client/src/services/serverAPI.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
 
-export const api = axios.create({
+export const serverAPI = axios.create({
   baseURL: 'http://localhost:3333',
 });
